+++ packages/util-javi/CHANGELOG.md
@@ -7,3 +7,4 @@
 ## :boom: Breaking changes
 
 - Moved `OpenFile` type to `jabc`.
+- Moved `algs` folder to new package `assorted-algorithms`.